ZIM 9.4.0 - Organizer, Panel, Pen and List Adjust
9.4.0
ORGANIZER
https://zimjs.com/explore/organizer.html
Added a ZIM Organizer class to organize a List
This lets the user add and remove items, and move them up and down in
the list
The Organizer can be found in the docs under components above Loader
near bottom of components
The organizer parameter of List can be used to add the organizer to the
list.
This will insert the organizer at the top of the list (and under the
titleBar if there is one)
Or a list can be passed to the Organizer's list parameter and then it
can be manually positioned
In making the GenPen app, the Organizer was used a half dozen times for
three different purposes - very handy

PANEL
https://zimjs.com/explore/panel.html
Added new Panel() class just before Pane() in Components.
This gives a box with a title bar and optional arrows for more than one
screen in panel
Panel is more simple than a Window - which has scrollbars, etc.
Panel can be dragged but has no close button - perhaps a Pane() would be
used for that
Passing in a series to the Label and color/backgroundColor will make
multiple screens (shown one at a time)
A choice of a single right arrow or left right arrows can be used to
cycle through the panels

PEN AND MOTIONCONTROLLER
Changed pressupStop parameter to immediateStop - (default true)
if pressup is used (drag or MotionController) then stop drawing
immediately when press is up
set to false to keep drawing until damping is finished.
Adjusted mouseDownIncludes on MotionController to include specified
object even when inside a container
Fixed bug in Pen with new paper after old paper has been moved
This was an addTo() - remember the default of addTo uses localToLocal to
match positions
Using shape.addTo(newPaper, 0, false) solved a very tricky bug... for
when you expect a 0,0 position.
Added "recordUndo" event to Pen to help tie pen in to outer undo system
Fixed up nib system - so now no invisible default nib - was causing
simultaneous draw and drag issues.
Nib with drag uses different placement than nib with MotionController,
etc.

LIST AND TABS - IMPROVEMENT
Added scrollBarOverlay parameter to List BREAK - after scrollBar
parameters
This defaults to true and will overlay the scrollBar on the list rather
than add space - which would leave a gap when scrollBar is not there
Adjusted List to not override padding with spacing.  So it will default
to spacing but you can alter padding, paddingVertical and
paddingHorizontal
Remove indentHorizontal and indentVertical - BREAK these were needed for
label indenting but then we added labelIndentHorizontal and
labelIndentVertical
The indent will work only on its direction - horizontal or vertical.
Padding works in all directions.
Adjusted the docs of List to describe spacing, indenting and padding
relationships.
Fixed selectedColor glitch in Tabs - was not starting with selectedColor
for text as this was pasted in the rollColor
Adjusted Tabs to use zim.copy with clone set to true for proper cloning.
Adjusted List and Toggle to use zim namespace in DUO function to work
properly with zns=true
Made vertical:false List slideSnap be "horizontal" by default
Added support for starting with an empty list
Adjusted Tabs and List removeFrom() to not give error if no items to
remove
Added original color and backgroundColor to custom Button objects passed
in to list or tabs
No longer giving "tabs have total less than width" message - instead
setting the width to the width after the tabs are made (BREAK)
Fixed Tabs backdropColor defaults - had typo for backgroundColor
defaults
Adjusted Tabs rollColor on first item
Added excludeCustomTap parameter to Tabs and List to let custom Button
objects have their own tap() - for instance, if previously set
Fixed glitch in height when using titleBar.
Made tap() ignore List titleBar and organizer
List now accepts corner in format of [topLeft, topRight, bottomRight,
bottomLeft]

TILE - IMPROVEMENT
Added ZIM VEE support to align and valign for Tile - now can set a
series() for different alignment!
Can already set ZIM VEE for different column and row sizes.
ZIM VEE for align and valign is not supported with matching squeezeH and
squeezeV specified - too complicated.

TAG - TEXTAREA
IMPROVEMENT These now support drag() and gesture() (no rotation) -
thanks KV for the catch!
The https://zimjs.com/explore/tag.html has been updated with a drag()
example at the bottom in red
backgroundColor, padding, paddingHorizontal, paddingVertical, expand
parameter have been added
The drag will naturally work around the edges of the HTML tag
The padding (default 10) will trigger drag as will the expand (default
20)
So if you have a large padding, you may want to set expand to 0 for
instance.
To make the whole tag drag use: myTag.style.pointerEvents = "none";

VEE
added vee(obj) to code module under featured
vee(obj) is a short function that returns true if obj is in ZIM VEE
format else false
ZIM VEE format is [], {min:a, max:b}, function(){}, {noZik:x}
ZIM VEE is a way to pass in dynamic parameters or style properties
This is very handy to pass in a series() function or an array for random
pickings, etc.
Used to create dynamic particles with the Emitter or tile specific items
in order, etc.
ZIM VEE accepts any value and if not in ZIM VEE format, will just return
the object
It is called ZIM VEE as it was launched with ZIM VEE (5) - we are on ZIM
NIO (9)

GENERAL
Fixed a glitch in MotionController - since Accelerator was added - was
not dispatching "moving" events properly
Adjusted loc() to not default to 0 if x or y is missing but rather keep
the object's current x and y
Fixed CheckBox style for checked style to work properly.
Fixed second arrows on Stepper to work since shifting parameters on
Triangle.
Made LabelOnPath accept a string for the label parameter - and then it
converts the string to a Label
Added automatic pointer to object with tap()
Fixed swatch text on ColorPicker to align "center" when changed