ZIM 9.4.1 - general updates to Pen, List, Panel...
9.4.1
LIST
Added an itemsText read-only property to List to give an array of item
label text - if no label then null in array.
Added a clear() method to the List
Fixed bug when enabling button having mouseChildren set to true - now
will remember this when re-enabling
By default, the Button has mouseChildren set to false but in some custom
buttons this is not desired.
This is rare - but the custom List buttons in GEN-PEN to drag, lock, and
hide were aversely affected.
Added getter setter method for text property of Label - so if Label is
manually changed, text property reflects change

PANEL
Added index parameter to nextPanel() on Panel - if using a series, can
go to specific index default null for next chosen panel.
Added event parameter to nextPanel() on Panel default false - set to
true to make nextPanel() trigger change event
Added an overlay property to Panel that makes the panel dark when
enabled is set to false
Added an extra parameter to Panel to give an extra button for reseting
for instance.

PEN
Added lineAlpha and lineBlendMode parameters and properties to Pen that
work on each line
Added sizeScale and spreadScale properties to Pen to scale these values
after picked from optional ZIM VEE values
Changed deleted Pen lines to be visible false rather than alpha false
(internal so new lineAlpha is easier)
Added Pen sizeFactor and spreadFactor properties that work independently
from sizeScale and spreadScale
This allows two dimensions of pen manipulation - for animating and
adjusting depth, for instance
Added note to Pen that cropScale should be kept at 1 for iOS to avoid
oversize canvas issues
Added lastSelected property to Pen that remembers the last selected
segment
Included pressmove events in enabled property of MotionController
Added infinite property to Pen to draw single segment until stop() is
called
Made animating pen along a path start with infinite and end with stop()
- also adjusted for pauseAnimate() and stopAnimate()

GENERAL
Added to Squiggle and Blob the following properties:
lastSelected - access to the last selected (or created) control
container
lastSelectedIndex - the index number of the last selected controls
Fixed a bug in animate that was not letting a path animate twice in some
cases - just set the target pathRatio to null when animate ends
pathRatio is used internally - percentComplete is used externally and
works when tween is running but was missing being able to set it ahead
of time...
Still may not set pathRatio ahead of time, but pathRatio is being set
back to null at end of animation and will be set to 0 at start of next
animation.
This may lead to a situation where we want to start at a
percentComplete...
oh... that should be okay as the animate will start at 0 and jump to the
percent complete on first Tick.
This is better than starting at 0 and jumping to the 1 (pathRatio max)
before percentComplete even kicks in...
Should be good... could possibly flash unless we caught that - will keep
an eye on it...
Fixed mousewheel scroll to work when hovering on Window (or List) - was
using e.stageY on a window.addEventListener (not going to work!)
Adjusted decimals time parameter to also show minutes : seconds rather
than minutes : decimal minutes
Added calculate() to MotionController Ticker to calculate speed on
manual x and y input
fixed Waiter dispose() to properly remove createjs tweens - thanks Frank
Los.
Added cursor to objects with hov() set to true
Added pausing the drag in animate() to the pause() method added to the
target by animate() - thanks Jaime Convery for the find
Limited ratio of ZIM colorRange() to between 0 and 1 - wiggle of color
sometimes goes a touch under or over resulting a hex number conversion
with one extra character (black)