ZIM 9.5.0 - Layer class, Emitter motion + fixes
9.5.0
LAYER
https://zimjs.com/explore/layer.html
Added a Layer component that provides a Container with transform
controls
ZIM transform() objects have their mousechildren turned off so they can
be dragged and transformed.
This means there can be no interactivity inside the transformed object.
Layer provides a solution to nest transformed objects in transformable
containers.
It does so by providing a titleBar that can be used to turn on and off
the transform of the container
and allow its contents to be transformed when the transform controls of
the Layer are turned off.
This is more than just hiding the transform tools but rather removing
and adding them.
Thank you, Stefan Soljemo for the suggestion on Slack
https://zimjs.com/slack

The Layer titleBar will always remain visible on the stage
If the Layer is moved (not by its titleBar) so that the titleBar hits
the edge,
then the titleBar will become anchored to the edge (unless anchor is set
to false)
This creates an independent titleBar that can be moved to any location.
The titleBarPos() method can also be used to separate the titleBar at
any time.
Drop the titleBar on the top left corner of the Layer or doubleClick it
to snap it back on to the layer

EMITTER
https://zimjs.com/2018/slack.html
Changed Emitter to emit to a particles container - this gets added right
under the emitter
This makes the emitter be moveable without moving all the particles that
have already been emitted.

KEYBOARD
Made Keyboard work with center and right aligned and scaled Labels as
well as backgroundColor and backing and padding.
Turned off resizing the keyboard on show(). You will need to custom
scale or position again after a keyboard.resize();  Thanks racheli golan

TRANSFORM
Added events parameter to transform() and to transformControls for
dynamic setting
This will turn on transformed event for pressmove on controls as well as
the traditional pressup.
Added a "transformshow" and "transformhide" events to add() and remove()
for objects with transform()

GENERAL
Added borderColor and borderWidth to RadioButtons and changed
backgroundColor to be background color of circle rather than border
color
Changed backgroundColor default to be white with .8 alpha
Fixed Sprite to accept CreateJS SpriteSheet - earlier testing for auto
image had disabled ZOE import.
Fixed Pane to receive a pattern - the rectangle had flatBottom param...
Adjusted Rectangle with and height defaults to be each other if only one
is provided - so providing one will make a square of that value
Added close and closeColor to styles for Pane
Added test in STYLE to not try and clone any stage or stageGL values.
Updated addTo, center, centerReg STYLE to allow single container value -
did require config objects (which still can be used).  They also accept
true
Button, List, Window, fixed cloning with border set to -1  - this was
being set to null and then cloned to be default of 1, etc.
Fixed rollColor property of Button - was setting label color not
rollColor
Fixed Button clone when border turned off and icon present - which
turned default to have a border