ZIM 10 - integrated physics, Pick, Hierarchy, etc.
PHYSICS
Integrated Physics into ZIM DisplayObjects and the Docs.
Still need to import Box2D and the ZIM physics module - updated to
version 2.
Making a physics object and mapping a ZIM object is no longer needed.
var circle = new Circle().center().addPhysics();
will create a default world and drop a circle in it!
Access the physics object with circle.physics
Or premake the world with new Physics() - to customize gravity, borders,
scroll, etc.
The addPhysics() lets you set a bunch of parameters about the body.
Access the physics body with circle.body

DisplayObject Methods:
A set of methods are added to the DisplayObject when physics is in
place:
impulse(), force() and torque() methods to push objects around and spin
them.
impulse() is a one time push like shooting a pool ball
force() is a force over time like gravity or wind and is applied in a
Ticker, etc.
torque() will rotate or spin the object around its center
The control() method will let keyboard keys move the object.
The follow() method will let the stage follow the object when the
Physics scroll parameter is set to true
Contact can be tested for with contact() and contactEnd()
each receives a callback function that is provided with the other
contacting body

Physics Methods:
The physics class has a join() method that can join objects in a variety
of physics joints.
The physics class has a drag() method to specify which objects are
draggable.
There is a debug() method to see the physics world behind ZIM.
The world can be set up to be bigger than the frame and scroll to follow
an object.

PICK
Added a Pick() class to featured code module
This takes the place of zik() when processing ZIM VEE values.
ZIM VEE is so handy that a more generic class was created to share with
the world of coding
Pick() accepts parameters for a series, an array for randomly picked, a
range object for a range or a function that returns a value
The Pick object has a choose() method that is used internally - but can
be used by coders in general.
There is also a chainable num() method that lets you set the number of
choices
The Pick object also has a loop() method so you can loop through the
options getting the next picked option each time.
If you have an array and want a series you can use JS6 Spread ... or use
new Options(series([your array]));
ZIM Pick is a stand-alone class that can be copied and used in other
languages for delayed parameters.

SVG CONTAINER
Added SVGContainer class - that turns SVG into ZIM Squiggles or Blob
paths
Thank you to KV from our Slack team https://zimjs.com/slack for the help
on this!
The tech is still in experimental stage but will translate most SVG - no
CSS on SVG...
The paths can then be used to animate along, drag along, make a
LabelOnPath, etc.
Also consider svgToBitmap() if you are just wanting to view an SVG file
in ZIM

HIERARCHY
Added a zim Hierarchy() class to accommodate nested data.
This is used by the accordion but could be used in making a "Mind Map"
or a Tree interface, etc.
The Hierarchy class comes with handy methods to traverse a hierarchy.
This is similar to DOM processing but uses Object literals and as such,
can store any Object

LIST ACCORDION
Added an accordion parameter to List to handle expanding and collapsing
sections to a List.
Uses ZIM Hierarchy for creating and tracking the nesting.

TIP
Added a Tip() class under components just after Toggle.
Tip() is a Label that has show() and hide() added to it
Tip will hide automatically 2000 ms after showing - or whatever you pass
in as a time.
Tip also has align, valign and margin, marginH, marginV to easily
position based on a target
There is also an outside parameter you can set to position outside a
target
Tip is added to the stage - initially it was added to the target but the
target might be rotated, etc.
Thanks Ami Hanya for the requestion for Toast ;-)

GENERAL
Added enabled property to List
Added selectedRollBackgroundColor and selectedRollColor to Tabs, List,
Pad - sigh <em>BREAK</em>.  These default to rollBackgroundColor and
rollColor
Added dampKeyup to MotionController and dampKeyup property to control
the damping on the Accelerator when keyup is used in MotionController.
See: https://zimjs.com/explore/sidescroller.html
Added expand to Window overlay rather than black with alpha .01
Made loop() work on an HTMLCollection (it already works on a NodeList)
Added q, r, s, t points to ZIM Point - used with SVG - Cubic Beziers and
Arc
Added addChar(char) and removeChar() to Keyboard()
Added x and y to rot() to rotate around any point (relative to the
container)
Added event object to hotSpot click - also added callOver and callOut
parameters
Added "myCanvasAlternative" system to show alternative content if no
canvas
or to screen readers and possibly search engine indexing
Added hotSpots property to HotSpots which is an array of HotSpot objects
Fixed sink to work with new Emitter system (particles container
independent of emitter position)
Adjusted Tile so it does not clone first item but rather uses item
passed in to Tile.
** patch
Added buttonDown property to Tabs and itemDown property to List to get
currently pressed item.
Added getItemIndex(item) method to List
Added clamp parameters to Proportion, ProportionDamp and Parallax
defaults to true
Set this to false to get values outside min and max range - thanks Ami
Hanya for the prompting
Adjusted scrollTop on Frame to default to false - this was for older
iPhones - live version 5 - and does nothing otherwise
It was reported that it might mess up mobile when tag mode is scrolled
to the bottom - so we have defaulted it to false.