ZIM 10 patches
Adjusted cloning - to initially only copy matrix if shape or not a ZIM
object
but realized that that lost cloning of x, y, scale and skew so backed
out of that change.
Added buttonDown property to Tabs and itemDown property to List to get
currently pressed item.
Added getItemIndex(item) method to List
Added clamp parameters to Proportion, ProportionDamp and Parallax
defaults to true
Set this to false to get values outside min and max range - thanks Ami
Hanya for the prompting
Adjusted scrollTop on Frame to default to false - this was for older
iPhones - live version 5 - and does nothing otherwise
It was reported that it might mess up mobile when tag mode is scrolled
to the bottom - so we have defaulted it to false.
Fixed stopAnimate() so it works with drag:true setting
Fixed Blob controlLength parameter - was not using parameter value - now
it is!
Made Squiggle and Blob move parameter only stop overall dragging when
set to false - that is what was intended
Passed a CreateJS tick object to any function added to Ticker with add()
this gives delta, time, timeStamp, etc. - thanks Alfred Yim for the
prompting!
Ticker.raw() functions receive a DOMHighResTimeStamp
Fixed Tile to properly not clone if clone:false is set - thanks Racheli
Golan for the find.
This was introduced initially in 10 for ZIM Shapes accepting ZIM VEE for
size and color - we mixed up a conditional and it is now patched.
Adjusted Blob to properly close path so there is no gap on thick borders
Controls were hiding behind transform objects since updates in 9.5
This has been fixed but should be monitored further.
NOTE - as of 2/26/2019
SVGContainer has been updated - thanks KV for all the testing and
updates on that.
Here is a mess of an example:
https://zimjs.com/explore/svgcontainer2.html