ZIM 10.0.1 - Circle Percent, Blob, Tile, STYLE...
PROTOTYPE CONTROL STYLE
Prototyping adding STYLE to controls - the first test is with Tile.
Added a delayPick style to let Tile choose the value rather than STYLE
on the object
For instance, passing [red, green, blue] to a Circle would pick the
color randomly
but adding a Circle to a tile gets only one styled circle - so the whole
tile would be that random colored circle
Adding delayPick:true to the circle will let the Control using the
object make the selection
This will be handy for Emitter, Pen, etc. when they get STYLE too.

Added a percent parameter to ZIM Circle that makes a Circle have a
percent arc.
So new Circle({percent:50}) is a semi-circle with bounds of the
semi-circle
The registration point stays the center of the circle
https://zimjs.com/explore/circleArcs.html

Added spin() method to Physics objects
this is a one-time turning force like an impulse is a one-time linear
force
To apply a constant turning force, use torque() in a Ticker or keydown
event, etc.
https://zimjs.com/explore/spin.html
Fixed Physics to work with scroll and follow using drag() - already was
working with controls but not drag.

Added pointsAdjusted property to Squiggle and Blob to change rotated or
scaled points to non-rotated and non-scaled
IMPROVEMENT Added normalize parameter to Squiggle and Blob update()
method to update to normalized points
must do this after rotating or scaling a point and then expecting manual
updates on points
otherwise if just animating rotation or scale of points do not need to
set normalize true (just leave parameter out)
Fixed bounds of cloned scaled Bitmap - was cloning width and height
rather than bounds.width and bounds.height - thanks Ami Hanya
Made Blob showControls param and method work when interactive false
IMPROVEMENT Fixed adding points to blob when showControls is false - it
remakes blob and was using showControls parameter value rather than
_controls dynamic value
Added faint background to custom Tabs and List elements so selected
index works for whole tab area not just directly on the custom item -
thanks Racheli Golan
https://zimjs.com/test/listAlign.html
Made Tabs and List align and valign work properly with objects that are
not top left registration point - and handles align of scaled custom
objects - thanks Ami Hanya
Fixed transformControls.update() to work after pos() - pos was using
CreateJS getTransformedBounds() and that was setting some sort of matrix
that mixed up transform
Fixed transformControls.update() to add controls on top of object - was
going one back since 9.5 update...
Fixed pressup on controls to remove custom cursors - lost a
stage.update() somewhen.
Adjusted Guide to show pixels on start when percent parameter is true
Fixed Sprite loopCount and call to work properly - animate() was reading
paused property of Sprite as true and not paying attention to loop count
Adjusted Accessibility html tags in behind to be rgba(0,0,0,.01) in
color to remain hidden when in tag scaling mode
Fixed Tile to work with negative scale objects this lets Tiles have rows
reversed or be fully reversed as follows:
https://zimjs.com/explore/tileflip.html