ZIM 10.3.0 - Retina, hitTestPath, Blob work...
RETINA
https://zimjs.com/retina.html
Added a retina parameter (default true) to scale the stage with
devicePixelRatio
Scaling the stage rather than the canvas works very well for scaling
vectors up
This is the technique that Adobe Animate uses - see the next section for
issues.

ADOBE ANIMATE
https://zimjs.com/animate.html
ZIM now provides a ZIM SHIM that lets ZIM work within Adobe Animate
without ZIM Frame.
This is a standard Animate Publishing template with a few lines of ZIM
code added.
ZIM was mostly supported in Animate but now, various issues below have
been solved.

Adobe Animate scales the stage - which affects globalToLocal,
localToGlobal, localToLocal.
ZIM now overrides these with fixes and dozens of changes have been made
as follows:

1. Basically localToGlobal needs its resulting point to be divided by
the stage scaleX and scaleY.
globalToLocal needs its input point multiplied by the stage scaleX and
scaleY
localToLocal needs to call these - but not the GlobalToLocal if the
object is on the stage directly - sigh.

2. Mouse even object e.stageX, e.stageY, e.rawX, e.rawY need to be
divided by the stage scaleX and scaleY.

3. stage.getObjectUnderPoint() needs to use the x and y multiplied by
stage scale
unless you are using e.stageX and e.stageY which would normally be
divided by stage scale
so the result is you can pass e.stageX and e.stageY directly in to
getObjectUnderPoint without any changes

4. Adjusted scaleTo() to assume stage is scale of 1 with stage.width and
stage.height matching what is set.

5. Shapes are still positioned as if on an unscaled stage for hitTests.
So any x and y values need to be multiplied by the stage scale before
doing a globalToLocal
into the shape to counter the globalToLocal expecting a 1 scale stage.

6. getConcatenatedMatrix() needs to have resulting values divided by
stage scale

OTHER ADOBE ISSUES

7. The CreateJS MovieClip does not have the ZIM fixes to globalToLocal,
localToGlobal and localToLocal
So create a ZIM Container with bounds of the MovieClip and add the
MovieClip to the container.
From then on, use the ZIM Container to drag, transform, hitTest, etc.

8. Adobe Animate is running an older version of CreateJS that does not
automatically recalculate (previously specified) bounds when set to
null.
This causes Tab heights to miscalculate which effects List, etc.
So adjusted Tabs to test for old CreateJS and not expect bound
recalculation.

9. There seems to be a problem with hitArea (expand()) on cached objects
so adjusted use one or the other

SQUIGGLE AND BLOB PATH
https://zimjs.com/hittestpath.html
More updates to path control to help out with Shape Tween which was
launched in 10.2.0:

1. Added addPoint(), addPoints() and interpolate() methods to squiggle
and blob
To add a point at a percentage, or add a number of points between points
Or get the data for any number of points - used by hitTestPoint
interpolate could be used to set dynamic bounds on Blob and Squiggle
which would lead to setting a drag Boundary - like the general drag()
has.
So look for this in future versions of ZIM.

2. We could shape tween between shapes with a different number of points
but for now, use the addPoint() manually so the number of points are
equal.
This will usually result in better shape tweens than trying to guess
where to automatically add a point.
But we may look into how that might be done via an algorithm.

3. Added hitTestPoint(num) to do a hitTest of any object shape along a
Squiggle or Blob path!

4. Adjusted the editPoints parameter so true adds points only to the
edge.
When we do, it adds the point and adjusts the points next to it - for no
overall path adjustment.
Thanks Sam Van Herck for the promptings.
If you want the original adding points anywhere - use
editPoints:"anywhere"
At which point, it adds a point with controlType "none".
Added two properties to adjust accuracy vs. speed of adding points:
addPointFactor - (default 20) used when placing new points along edge
(editPoints is true)
divides the distance between points by this amount - the smaller the
more accurate but also slower
addMinDistance - (default 15) edge press needs to be within this
distance to add a point to the edge

GENERAL
Adjusted Stage and StageGL to accept a canvas tag as well as the
traditional canvas tag ID.
Added rgb and rgba support to ColorPicker - thanks Ami Hanya for the
promptings
The rgba alpha now adds a start alpha for startBackgroundColor - if rgba
is used in a color list the alpha is ignored
Added rectIntersect(a,b,margin) function to code module which returns
true if rectangles intersect
Added boundsAroundPoints(points) function to code module to get a
rectangle around an array of points
These are used by Blob and Squiggle and hitTestPath
Fixed Blob "circle", "rectangle", "triangle" points parameter values to
use radius setting as well.
Fixed "rectangle" to be centerReg() for Blob IMPROVEMENT
Fixed slight spacing issue with hitTestGrid() IMPROVEMENT
Undecided how to treat setting bounds width or height to 0
it means that width and height can't be set as they adjust the scale and
we divide by 0
So adjusted Tabs to overwrite the width if width is 0
Added blur event to Window to stop drag if browser window loses focus -
thanks Ami Hanya
Seems that maybe mobile was not handling transform controls properly due
to getObjectUnderPoint test
So removed this requirement for mobile - means objects overlayed on
transforms may trigger transforms on
This was a rare issue and more of a perfectionist fix...
The getObjectUnderPoint is working in a simple case on mobile but not in
the complexities of transform().