ZIM 10.4.0 - List CheckBox, Path traverse, fixes...
Added checkBox parameter to List to use checkboxes in list - thanks
Dale789 for the prompting!
Added checkBoxes property to List to give a read only array of
checkBoxes
Added a checkBox property to a list item like list.selected.checkBox to
give access to an item checkBox
Added the following three methods to manipulate checkBoxes:
setCheck(index, type) - set the CheckBox at an index to true or false
(the type parameter)
setChecks(type) - set all CheckBoxes to true or false (the type
parameter) returns object for chaining
getCheck(index) - get the checked value of the CheckBox at an index
Added tap parameter to CheckBox and Toggle to activate on tap - good for
List and Window when dragging
Added toggle(type) to CheckBox and toggled property - same as setChecked
and checked but consistent with others

https://zimjs.com/ten/traverse.html
Added a traverse() method to Blob and Squiggle to animate an object
between points of the path
Added a "traversed" event that is dispatched when a traverse() ends
To traverse backwards past start or forward past end of Blob use two
traverse calls
Thanks KV for the experimentation and requests.
<strong>IMPROVEMENT</strong> Fixed up animate() along path so it
properly resets internal pathRatio and tween with stopAnimate() called
this is important when doing multiple animations along path.
Set this to reset only with path or dynamic animation - watch out for
color animation
Color seems to work but may have issues with multiple color animations
on same object

ZIM three_2.0.js has launched to work with retina
The namespace is now optional and works like new Three();
See https://codepen.io/zimjs/pen/qGPVqO

Adjusted pizzazz_03 to work with retina
https://zimjs.com/patterns.html
Tidied up setMask() for retina - was not working if mask had same parent
as object.
This fixed the ProgressBar() in retina
This is an odd localToLocal issue since retina - we will keep an eye on
it.

GENERAL
Fixed zta() so it does not break in IE11 - as there is no
console.table()
Patched animate() in series mode to make sure last animate object call
is called - thanks Yalon for the find