ZIM 10.5.0 - Marquee, List pulldown, animate fixes
CDN UPDATE
https://zimjs.org/cdn/
Changed the CDN to zimjs.org/cdn/ with a new naming format
The Amazon d309knd7es5f10.cloudfront.net url still works for OLDER
versions of ZIM
The new naming format uses the directory to keep track of ZIM versions
The file names will remain constant as zim.js for minified and
zim_doc.js for full
So the two main libraries are at:
https://zimjs.org/cdn/createjs_1.2_min.js
https://zimjs.org/cdn/10.5.0/zim.js
The latest helper libraries have been moved to the new CDN as well
See the top of the Docs pages for links
Explanation: as we move into more than a million views a week amazon was
costing money
zimjs.org uses CloudFlare which caches the files at the front edge of
the internet cloud!
Many other js libraries use CloudFlare too at https://cloudflare.com

MARQUEE
https://zimjs.com/marquee.html
Added a Marquee component to cycle through images and interactive works
Uses ZIM Pages and Indicator objects - and a pause/play button
Can show ZIM interactive works in banner area!
Docs can be found near the bottom of Components

LIST PULLDOWN
https://zimjs.com/ten/pulldown.html
Added pulldown parameter to List
which prepares the list to take a height based on a number of items
It also hides the background and border
Setting this with accordion can expand and contract the list like a
pulldown
Added open property to List accordion object that starts an accordion
open

LIST ITEMS
https://zimjs.com/explore/pulldown.html
Created special List items: Slider, CheckBox, ColorPicker (for now)
These are created as List.slider(), List.checkBox(), List.colorPicker()
and can be combined with the pulldown arrangement to make a collapsible
control panel

DOT ANIMATE
Updated CreateJS version to include the new TweenJS dot plugin
https://zimjs.org/cdn/createjs_1.2_min.js
This allows animation of properties within properties by using quoted
values for the property name
so props:{"rotation.x":200} for instance
Updated dozens of property assignments within animate() to handle dot
property values
Note, TweenJS requires an initial . at the start
props:{".rotation.x":200}
but that is not easy to remember so ZIM allows for both notations.
TweenJS also added a plugin for relative animation using "" - ZIM
already has that
the TweenJS version requires a + or - at the start.  ZIM does not
require the +
Unfortunately, the dot plugin requires the TweenJS version
So we have made ZIM convert any relative without a + to having a +
which means, we are good to go with or without the + for relative
animation
props:{".rotation.x":"30"} or props:{".rotation.x":"+30"} will both work

REPLAY TWEEN
Added resetTween() and replayTween() methods to objects with tweens
These will set the object to the tweened property values before the
tween started
and replay the tween - re-runs the animate() with the same parameters as
the last tween
Added a clean parameter to animate() that will not clear the tween ids
when tween ends
the idea being percentComplete will still work - but unfortunately, it
does not include the wait time
so replayTween() was added if the initial wait time is desired

GENERAL
Added makeIcon() to Frame to make the ZIM TEN Z icon (360 byte
customizable vector)
Added animateReturn() method on target of animate()
that works when animating on path to return target to the start if
rewind is set to true
IMPROVEMENT - fixed Blob and Squiggle due to change in center() and
centerReg() - was shifting points
IMPROVEMENT - Fixed Slider to capture stage variable when track is
pressed - was needing button pressed first since Retina
Fixed negative sign on list hierarchy - it was shifting too low on
retina
Fixed Dynamo and Accelerator pause to not break if paused the same way
twice - it now ignores the second time
Made Dynamo remember change in percentSpeed when paused so unpausing is
at new speed
Added read/write sensitivity property to ZIM Swiper
Added conversion from rgb() and rgba() to hex and string to
convertColor() - thanks Ikki for reminding us!
also made conversion to string output hex number if no matching string
color
Fixed Label center on backing - had removed code there for some reason
in 10.3.0 - look back and see why
Also tidied up bounds of label with scaled backing - thanks Bart Libert
for the report
Fixed hitTestPoint, hitTestReg, hitTestRect and hitTestCircles to work
properly with Retina - thanks again Bart Libert
Made ColorPicker force upperCase for selectedColor
Fixed List indenting issue with align left or right with both custom and
label items
Added checkBox parameter to styles for List
Adjusted animate() sequence to work properly with pauseAnimate()
previously, the sequence was run by timeouts initially that were not
part of the pauseAnimate system
so startPaused:true for instance would let the timeouts run
and the sequence would run all at the same time when pauseAnimate(false)
was called.
So changed sequence to use wait animation and then wait again if a wait
was in the sequence
this ties sequence into the animation chain from the start
IMPROVEMENT Adjusted ZIM Label to better work with ZIM Retina.
The stage scale was affecting the placement of createjs and perhaps
canvas text relative to bounds
Reworked how the Label centers and positions as well as on backing and
backgrounds
Changed Tabs and Stepper to center label
Adjusted docs on Label to read the correct "top" as valign default
ZIM strays from HTML tables on this with a default left and top
alignment for text
However for Buttons, etc. the default is center
Buttons support ZIM VEE (Pick / zik) on main colors - see docs
This was primarily for alternating colors on Lists for instance