10.5.2 - Frame.follow(), hitTest fixes, series...
FOLLOW
https://zimjs.com/ten/follow.html - with keys
https://codepen.io/danzen/pen/gNKQYY - with press
Added follow() method to ZIM Frame - works like follow() in physics -
thanks John Smith for the request
but can follow any DisplayObject inside a Container - not just on stage
Added a followBoundary property to Frame to update the follow boundary
in a "resize" event if frame is in "full" mode
With physics we were generally forced to make follow move the stage
but here we can add the object to a container and make the container
move rather than the stage
this is a little more conventional as traditionally, the stage does not
move.
Added a motionController property to any target being controlled by a
ZIM MotionController
this is used internally by ZIM for follow - but could be generally
useful.
For ZIM MotionController type parameter, added a "follow" that keeps obj
moving towards pressed position

GENERAL
IMPROVEMENT - adjusted ZIM MotionController so does not test for object
beneath mouse if there are no mouseDownIncludes
this was lagging when there are thousands of objects on the stage and
was unnecessarily
Adjusted hitTests for Point, Reg, Rect, Circle and Path to use a
slightly different calculation
for Bitmaps and Sprites versus all other DisplayObjects - IMPROVEMENT
The latter divides by stage scale where the former does not. (patched
10.5.0 and 10.5.1)
Changed a few timeout() and mobile() calls to zim.timeout() and
zim.mobile() for zns=true setting - thanks Alfred Yim for the report
Fixed TransformManager and Transform to properly record last selected if
not moved and not flash transform (since 9.5.0 Layer code)
Tested to make sure there is a titleBar before setting titleBarPosition
in Layer
IMPROVEMENT - fixed ZIM animate() series so relative setting uses
calculated values
Was using last actual target value and this was compounding a slight
timing/value error
Now, for instance, if wait is added to series, resting place matches
desired values exactly
see https://zimjs.com/codepen/triangles.html and source for comments...
Patched physics_2.0.js to include a frame property (thanks James Barrett
for the request)
now physics can be transferred from frame to frame IMPROVEMENT
the drag() method will need to be reset each time
See https://zimjs.com/expand/physics.html for an example
where we go from a tag mode to a fit mode and change the frame of the
physics object